Contacted Regions Added to TraceSummary (#3353)
* Contacted Regions Added to TraceSummary

* Comments resolved

* Contracts Updated

* Contracts Updated

* Contracts Updated

* Contacted Regions removed from constructor

* TraceSummaryissues solved

* Summary Tests modified for contacted regiosn
