+++ Microsoft.Azure.Cosmos/src/Diagnostics/CosmosTraceDiagnostics.cs
@@ -46,7 +46,7 @@ namespace Microsoft.Azure.Cosmos.Diagnostics
 
         public override IReadOnlyList<(string regionName, Uri uri)> GetContactedRegions()
         {
-            return this.Value?.RegionsContacted;
+            return this.Value?.Summary?.RegionsContacted;
         }
 
         internal bool IsGoneExceptionHit()
