+++ Microsoft.Azure.Cosmos/src/Tracing/Trace.cs
@@ -17,7 +17,6 @@ namespace Microsoft.Azure.Cosmos.Tracing
         private static readonly IReadOnlyDictionary<string, object> EmptyDictionary = new Dictionary<string, object>();
         private readonly List<ITrace> children;
         private readonly Lazy<Dictionary<string, object>> data;
-        private readonly ISet<(string, Uri)> regionContactedInternal;
         private ValueStopwatch stopwatch;
 
         private Trace(
@@ -25,7 +24,6 @@ namespace Microsoft.Azure.Cosmos.Tracing
             TraceLevel level,
             TraceComponent component,
             Trace parent,
-            ISet<(string, Uri)> regionContactedInternal,
             TraceSummary summary)
         {
             this.Name = name ?? throw new ArgumentNullException(nameof(name));
@@ -37,8 +35,6 @@ namespace Microsoft.Azure.Cosmos.Tracing
             this.Parent = parent;
             this.children = new List<ITrace>();
             this.data = new Lazy<Dictionary<string, object>>();
-
-            this.regionContactedInternal = regionContactedInternal;
             this.Summary = summary ?? throw new ArgumentNullException(nameof(summary));
         }
 
@@ -61,38 +57,6 @@ namespace Microsoft.Azure.Cosmos.Tracing
         public IReadOnlyList<ITrace> Children => this.children;
 
         public IReadOnlyDictionary<string, object> Data => this.data.IsValueCreated ? this.data.Value : Trace.EmptyDictionary;
-
-        public IReadOnlyList<(string, Uri)> RegionsContacted 
-        { 
-            get
-            {
-                lock (this.regionContactedInternal)
-                {
-                    return this.regionContactedInternal.ToList();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Update region contacted information to this node
-        /// </summary>
-        /// <param name="traceDatum"></param>
-        public void UpdateRegionContacted(TraceDatum traceDatum)
-        {
-            if (traceDatum is ClientSideRequestStatisticsTraceDatum clientSideRequestStatisticsTraceDatum)
-            {
-                if (clientSideRequestStatisticsTraceDatum.RegionsContacted == null || 
-                            clientSideRequestStatisticsTraceDatum.RegionsContacted.Count == 0)
-                {
-                    return;
-                }
-               
-                lock (this.regionContactedInternal)
-                {
-                    this.regionContactedInternal.UnionWith(clientSideRequestStatisticsTraceDatum.RegionsContacted);
-                }
-            }
-        }
 
         public void Dispose()
         {
@@ -123,7 +87,6 @@ namespace Microsoft.Azure.Cosmos.Tracing
                 level: level,
                 component: component,
                 parent: this,
-                regionContactedInternal: this.regionContactedInternal,
                 summary: this.Summary);
 
             this.AddChild(child);
@@ -156,14 +119,13 @@ namespace Microsoft.Azure.Cosmos.Tracing
                 level: level,
                 component: component,
                 parent: null,
-                regionContactedInternal: new HashSet<(string, Uri)>(),
                 summary: new TraceSummary());
         }
 
         public void AddDatum(string key, TraceDatum traceDatum)
         {
             this.data.Value.Add(key, traceDatum);
-            this.UpdateRegionContacted(traceDatum);
+            this.Summary.UpdateRegionContacted(traceDatum);
         }
 
         public void AddDatum(string key, object value)
