+++ Microsoft.Azure.Cosmos/src/Tracing/TraceSummary.cs
@@ -6,8 +6,10 @@ namespace Microsoft.Azure.Cosmos.Tracing
 {
     using System;
     using System.Collections.Generic;
+    using System.Linq;
     using System.Text;
     using System.Threading;
+    using Microsoft.Azure.Cosmos.Tracing.TraceData;
 
     /// <summary>
     /// The total count of failed requests for an <see cref="ITrace"/>.
@@ -39,6 +41,59 @@ namespace Microsoft.Azure.Cosmos.Tracing
         public int GetFailedCount()
         {
             return this.failedRequestCount;
+        }
+
+        /// <summary>
+        /// Consolidated Region contacted Information of this and children nodes
+        /// </summary>
+        private readonly HashSet<(string, Uri)> regionContactedInternal = new HashSet<(string, Uri)>();
+
+        /// <summary>
+        /// Consolidated Region contacted Information of this and children nodes
+        /// </summary>
+        public IReadOnlyList<(string, Uri)> RegionsContacted
+        {
+            get
+            {
+                lock (this.regionContactedInternal)
+                {
+                    return this.regionContactedInternal.ToList();
+                }
+            }
+        }
+
+        /// <summary>
+        /// Update region contacted information to this node
+        /// </summary>
+        /// <param name="traceDatum"></param>
+        public void UpdateRegionContacted(TraceDatum traceDatum)
+        {
+            if (traceDatum is ClientSideRequestStatisticsTraceDatum clientSideRequestStatisticsTraceDatum)
+            {
+                if (clientSideRequestStatisticsTraceDatum.RegionsContacted == null ||
+                            clientSideRequestStatisticsTraceDatum.RegionsContacted.Count == 0)
+                {
+                    return;
+                }
+
+                lock (this.regionContactedInternal)
+                {
+                    this.regionContactedInternal.UnionWith(clientSideRequestStatisticsTraceDatum.RegionsContacted);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Add region contacted information to this node
+        /// </summary>
+        /// <param name="regionName"></param>
+        /// <param name="locationEndpoint"></param>
+        public void AddRegionContacted(string regionName, Uri locationEndpoint)
+        {
+            lock (this.regionContactedInternal)
+            {
+                this.regionContactedInternal.Add((regionName, locationEndpoint));
+            }
         }
 
     }
